+++ src/features/cockpit.js
@@ -366,7 +366,7 @@ class FTraceViewDataProvider extends BaseDataProvider {
         this._onDidChangeTreeData = new vscode.EventEmitter();
         this.onDidChangeTreeData = this._onDidChangeTreeData.event;
 
-        this.data = null;
+        this.data = null;  //json {item: {a:object, b:object, c:object}}
         this.documentUri = null;
     }
 
@@ -374,20 +374,45 @@ class FTraceViewDataProvider extends BaseDataProvider {
         if(this.data === null || this.documentUri === null){
             return [];
         }
-        return this.data.map(k => {
+        return Object.keys(this.data).map(k => {
+            let children = this.data[k];
             return { 
+                children: children,
                 resource: this.documentUri, //uri
                 label: k,
                 tooltip: k,
                 name: k, 
                 parent: null,
                 iconPath: vscode.ThemeIcon.File,
+                collapsibleState: children && Object.keys(children).length > 0 ? vscode.TreeItemCollapsibleState.Expanded : 0,
             };
         });
     }
 
-    dataGetChildren(){
-        return null; //no children :)
+    dataGetChildren(element){
+        if(!element) {
+            return this.data;
+        }
+
+        if(!element.children){
+            return [];
+        }
+        // element provided? - 
+        return Object.keys(element.children).map(k => {
+            return { 
+                resource: this.documentUri, //uri
+                label: k,
+                tooltip: k,
+                name: k, 
+                parent: element,
+                iconPath: vscode.ThemeIcon.File,
+            };
+        });
+    }
+
+
+    dataGetParent(element){
+        return element.parent;
     }
 
     /** events */
@@ -456,10 +481,9 @@ class FTraceView extends BaseView {
             functionName = "<Fallback>";
         }
 
-        let ret = surya.ftrace(contractName + "::" + functionName, 'all', files, {}, true).trim();
-        let lines = ret.match(/^.*((\r\n|\n|\r)|$)/gm);
+        let retj = surya.ftrace(contractName + "::" + functionName, 'all', files, {jsonOutput: true}, true);
         this.dataProvider.documentUri = documentUri;
-        this.dataProvider.data = lines;
+        this.dataProvider.data = retj;
         this.refresh();
     }
 }
@@ -467,7 +491,17 @@ class FTraceView extends BaseView {
 /* Methods View */
 
 
-class PublicMethodsViewDataProvider extends FTraceViewDataProvider {
+class PublicMethodsViewDataProvider extends BaseDataProvider {
+
+    constructor(treeView){
+        super();
+        this.treeView = treeView;
+        this._onDidChangeTreeData = new vscode.EventEmitter();
+        this.onDidChangeTreeData = this._onDidChangeTreeData.event;
+
+        this.data = null;
+        this.documentUri = null;
+    }
 
     async dataGetRoot(){
         if(this.data === null || this.documentUri === null){