+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 All notable changes to the "solidity-visual-auditor" extension will be documented in this file.
 
 ## v0.0.23
-- new: Update notifications! - can be disabled of course :) 
+- new: Update notifications have arrived!
+- updated: solidity parser and surya
+- new: 🔥 Solidity Visual Auditor Cockpit panel
+    - Workspace Explorer
+    - Quick-access to extension settings
+    - Find Top Level Contracts
+    - Keep track of flattened files
+    - List public state-changing methods from the current contract
+    - Show the function call trace for the current method
 
 ## v0.0.22
 - update: solidity parser, surya (#41 [#42](https://github.com/tintinweb/vscode-solidity-auditor/issues/42))